+++ lib/verikloak-bff.rb
@@ -18,3 +18,6 @@ require 'verikloak/bff/proxy_trust'
 require 'verikloak/bff/forwarded_token'
 require 'verikloak/bff/consistency_checks'
 require 'verikloak/bff/header_guard'
+require 'verikloak/bff/rails'
+
+require 'verikloak/bff/railtie' if defined?(Rails::Railtie)